face/docker-compose/face-service/requirements.txt
fastapi
uvicorn
numpy
opencv-python-headless
insightface
onnxruntime
faiss-cpu
pymysql
python-multipart